+++ lib/terradactyl/terraform/commands/init.rb
@@ -59,6 +59,12 @@ module Terradactyl
       end
     end
 
+    module Rev1_02
+      module Init
+        include Terradactyl::Terraform::Rev015::Init
+      end
+    end
+
     module Commands
       class Init < Base
       end