+++ lib/terradactyl/terraform/planfile.rb
@@ -108,6 +108,11 @@ module Terradactyl
       end
     end
 
+    module Rev1_02
+      class PlanFileParser < Rev012::PlanFileParser
+      end
+    end
+
     module Rev011
       class PlanFileParser < Rev012::PlanFileParser
         def checksum