+++ spec/fixtures/stacks/rev1_02/test.tf
@@ -1,0 +1,6 @@
+provider "null" {
+}
+
+resource "null_resource" "foo" {
+}
+